+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -10,8 +10,8 @@
  */
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	int val = n;
-	int mask = 1 << val;
+	unsigned long int val = n;
+	unsigned long int mask = 1 << val;
 	if ((val & mask) | (0 << index))
 		return (1);
 	return (-1);
